+++ src/app/services/sleep-detection-automations/sleep-mode-for-sleep-detector-automation.service.ts
@@ -6,7 +6,7 @@ import {
   SleepModeEnableForSleepDetectorAutomationConfig,
 } from '../../models/automations';
 import { cloneDeep } from 'lodash';
-import { BehaviorSubject, firstValueFrom, Observable } from 'rxjs';
+import { BehaviorSubject, distinctUntilChanged, firstValueFrom, Observable, skip } from 'rxjs';
 import { SleepService } from '../sleep.service';
 import { SleepDetectorStateReport } from '../../models/events';
 import { NotificationService } from '../notification.service';
@@ -19,6 +19,7 @@ import { EventLogService } from '../event-log.service';
 export class SleepModeForSleepDetectorAutomationService {
   private sleepEnableTimeoutId: number | null = null;
   private lastEnableAttempt = 0;
+  private lastSleepModeDisable = 0;
   private enableConfig: SleepModeEnableForSleepDetectorAutomationConfig = cloneDeep(
     AUTOMATION_CONFIGS_DEFAULT.SLEEP_MODE_ENABLE_FOR_SLEEP_DETECTOR
   );
@@ -48,6 +49,9 @@ export class SleepModeForSleepDetectorAutomationService {
     this.automationConfig.configs.subscribe(
       (configs) => (this.enableConfig = configs.SLEEP_MODE_ENABLE_FOR_SLEEP_DETECTOR)
     );
+    this.sleep.mode.pipe(distinctUntilChanged(), skip(1)).subscribe((mode) => {
+      if (!mode) this.lastSleepModeDisable = Date.now();
+    });
     await listen<SleepDetectorStateReport>('SLEEP_DETECTOR_STATE_REPORT', (event) =>
       this.handleStateReportForEnable(event.payload)
     );
@@ -84,6 +88,8 @@ export class SleepModeForSleepDetectorAutomationService {
       return;
     // Stop here if the last time we tried enabling was less than 15 minutes ago
     if (Date.now() - this.lastEnableAttempt < 1000 * 60 * 15) return;
+    // Stop here if the last time we disabled sleep mode was less than 15 minutes ago
+    if (Date.now() - this.lastSleepModeDisable < 1000 * 60 * 15) return;
     // Attempt enabling sleep mode
     this.lastEnableAttempt = Date.now();
     // If necessary, first check if the user is asleep, allowing them to cancel.